MGMT-19876: Move upgrade functions to upgrade package
Move IsUpgradeRequested and GetWorkloadClusterVersion to
their own package to keep the controller package slimmer.
