+++ controlplane/internal/controller/openshiftassistedcontrolplane_controller.go
@@ -19,9 +19,9 @@ package controller
 import (
 	"context"
 	"errors"
-	"fmt"
 	"time"
 
+	"github.com/openshift-assisted/cluster-api-agent/controlplane/internal/upgrade"
 	"github.com/openshift-assisted/cluster-api-agent/controlplane/internal/version"
 	"github.com/openshift-assisted/cluster-api-agent/controlplane/internal/workloadclient"
 
@@ -32,7 +32,6 @@ import (
 	"github.com/openshift-assisted/cluster-api-agent/controlplane/internal/auth"
 	"github.com/openshift-assisted/cluster-api-agent/util"
 	logutil "github.com/openshift-assisted/cluster-api-agent/util/log"
-	configv1 "github.com/openshift/api/config/v1"
 	hivev1 "github.com/openshift/hive/apis/hive/v1"
 
 	corev1 "k8s.io/api/core/v1"
@@ -186,11 +185,11 @@ func (r *OpenshiftAssistedControlPlaneReconciler) Reconcile(ctx context.Context,
 	k8sVersion, err := r.OpenShiftVersion.GetK8sVersionFromReleaseImage(ctx, releaseImage, oacp)
 	markKubernetesVersionCondition(oacp, err)
 	oacp.Status.Version = k8sVersion
-	if oacp.Status.DistributionVersion, err = GetWorkloadClusterVersion(ctx, r.Client, r.WorkloadClusterClientGenerator, oacp); err != nil {
+	if oacp.Status.DistributionVersion, err = upgrade.GetWorkloadClusterVersion(ctx, r.Client, r.WorkloadClusterClientGenerator, oacp); err != nil {
 		log.Error(err, "failed to set the openshift version in the control plane status")
 	}
 
-	if IsUpgradeRequested(ctx, oacp) {
+	if upgrade.IsUpgradeRequested(ctx, oacp) {
 		//TODO: Handle upgrade request
 		log.Info("upgrade for control plane has been requested")
 	}
@@ -210,32 +209,6 @@ func markKubernetesVersionCondition(oacp *controlplanev1alpha2.OpenshiftAssisted
 	} else {
 		conditions.MarkTrue(oacp, controlplanev1alpha2.KubernetesVersionAvailableCondition)
 	}
-}
-
-func IsUpgradeRequested(ctx context.Context, oacp *controlplanev1alpha2.OpenshiftAssistedControlPlane) bool {
-	log := ctrl.LoggerFrom(ctx)
-	oacpDistVersion, err := semver.NewVersion(oacp.Spec.DistributionVersion)
-	if err != nil {
-		log.Error(err, "failed to detect OpenShift version from ACP spec", "version", oacp.Spec.DistributionVersion)
-		return false
-	}
-
-	upgrade := false
-	if oacp.Status.DistributionVersion == "" {
-		return false
-	}
-	currentOACPDistVersion, err := semver.NewVersion(oacp.Status.DistributionVersion)
-	if err != nil {
-		log.Error(err, "failed to detect OpenShift version from ACP status", "version", oacp.Spec.DistributionVersion)
-		return false
-	}
-
-	if oacpDistVersion.Compare(*currentOACPDistVersion) > 0 {
-		log.Info("Upgrade detected, new requested version is greater than current version",
-			"new requested version", oacpDistVersion.String(), "current version", currentOACPDistVersion.String())
-		upgrade = true
-	}
-	return upgrade
 }
 
 // Ensures dependencies are deleted before allowing the OpenshiftAssistedControlPlane to be deleted
@@ -529,62 +502,4 @@ func (r *OpenshiftAssistedControlPlaneReconciler) ensurePullSecret(
 	}
 	acp.Spec.Config.PullSecretRef = &corev1.LocalObjectReference{Name: secret.Name}
 	return nil
-}
-
-func GetWorkloadClusterVersion(ctx context.Context, client client.Client,
-	workloadClusterClientGenerator workloadclient.ClientGenerator,
-	oacp *controlplanev1alpha2.OpenshiftAssistedControlPlane) (string, error) {
-	workloadClient, err := getWorkloadClient(ctx, client, workloadClusterClientGenerator, oacp)
-	if err != nil {
-		return "", err
-	}
-
-	var clusterVersion configv1.ClusterVersion
-	if err := workloadClient.Get(ctx, types.NamespacedName{Name: "version"}, &clusterVersion); err != nil {
-		err = errors.Join(err, fmt.Errorf(("failed to get ClusterVersion from workload cluster")))
-		return "", err
-	}
-
-	return clusterVersion.Status.Desired.Version, nil
-}
-
-func getWorkloadClient(ctx context.Context, client client.Client,
-	workloadClusterClientGenerator workloadclient.ClientGenerator,
-	oacp *controlplanev1alpha2.OpenshiftAssistedControlPlane) (client.Client, error) {
-	if !isKubeconfigAvailable(oacp) {
-		return nil, fmt.Errorf("kubeconfig for workload cluster is not available yet")
-	}
-
-	kubeconfigSecret, err := util.GetClusterKubeconfigSecret(ctx, client, oacp.Labels[clusterv1.ClusterNameLabel], oacp.Namespace)
-	if err != nil {
-		err = errors.Join(err, fmt.Errorf("failed to get cluster kubeconfig secret"))
-		return nil, err
-	}
-
-	if kubeconfigSecret == nil {
-		return nil, fmt.Errorf("kubeconfig secret was not found")
-	}
-
-	kubeconfig, err := util.ExtractKubeconfigFromSecret(kubeconfigSecret, "value")
-	if err != nil {
-		err = errors.Join(err, fmt.Errorf("failed to extract kubeconfig from secret %s", kubeconfigSecret.Name))
-		return nil, err
-	}
-
-	workloadClient, err := workloadClusterClientGenerator.GetWorkloadClusterClient(kubeconfig)
-	if err != nil {
-		err = errors.Join(err, fmt.Errorf("failed to establish client for workload cluster from kubeconfig"))
-		return nil, err
-	}
-	return workloadClient, nil
-}
-
-// isKubeconfigAvailable returns true if the openshift assisted control plane
-// condition KubeconfigAvailable is true
-func isKubeconfigAvailable(oacp *controlplanev1alpha2.OpenshiftAssistedControlPlane) bool {
-	kubeconfigFoundCondition := util.FindStatusCondition(oacp.Status.Conditions, controlplanev1alpha2.KubeconfigAvailableCondition)
-	if kubeconfigFoundCondition == nil {
-		return false
-	}
-	return kubeconfigFoundCondition.Status == corev1.ConditionTrue
 }
